+++ src/index.ts
@@ -26,14 +26,12 @@ export {
   getSourceIri,
   getContentType,
   getResourceInfo,
-  getResourceInfoWithAcl,
   getPodOwner,
   isPodOwner,
   FetchError,
 } from "./resource/resource";
 export {
   getFile,
-  getFileWithAcl,
   deleteFile,
   saveFileInContainer,
   overwriteFile,
@@ -47,7 +45,6 @@ export {
   saveSolidDatasetInContainer,
   createContainerInContainer,
   deleteContainer,
-  getSolidDatasetWithAcl,
   solidDatasetAsMarkdown,
   changeLogAsMarkdown,
 } from "./resource/solidDataset";
@@ -140,10 +137,21 @@ export {
   getFallbackAcl,
   hasResourceAcl,
   getResourceAcl,
+  getSolidDatasetWithAcl,
+  getFileWithAcl,
+  getResourceInfoWithAcl,
   createAcl,
   createAclFromFallbackAcl,
   saveAclFor,
   deleteAclFor,
+  hasAccessibleAcl,
+  WithAccessibleAcl,
+  WithAcl,
+  WithFallbackAcl,
+  WithResourceAcl,
+  AclDataset,
+  AclRule as internal_AclRule,
+  Access,
 } from "./acl/acl";
 export {
   AgentAccess,
@@ -187,14 +195,6 @@ export {
   hasServerResourceInfo,
   WithResourceInfo,
   WithChangeLog,
-  hasAccessibleAcl,
-  WithAccessibleAcl,
-  WithAcl,
-  WithFallbackAcl,
-  WithResourceAcl,
-  AclDataset,
-  AclRule as internal_AclRule,
-  Access,
   UploadRequestInit,
   SolidClientError,
 } from "./interfaces";
