+++ src/resource/resource.internal.ts
@@ -20,52 +20,9 @@
  */
 
 import LinkHeader from "http-link-header";
-import {
-  WithAcl,
-  hasAccessibleAcl,
-  WithServerResourceInfo,
-  Access,
-  SolidDataset,
-} from "../interfaces";
-import {
-  internal_fetchResourceAcl,
-  internal_fetchFallbackAcl,
-} from "../acl/acl.internal";
-import { internal_defaultFetchOptions } from "./resource";
+import { Access } from "../acl/acl";
+import { WithServerResourceInfo, SolidDataset } from "../interfaces";
 import { clone as cloneDataset } from "../rdfjs";
-
-/**
- * This (currently internal) function fetches the ACL indicated in the [[WithServerResourceInfo]]
- * attached to a resource.
- *
- * @internal
- * @param resourceInfo The Resource info with the ACL URL
- * @param options Optional parameter `options.fetch`: An alternative `fetch` function to make the HTTP request, compatible with the browser-native [fetch API](https://developer.mozilla.org/en-US/docs/Web/API/WindowOrWorkerGlobalScope/fetch#Parameters).
- */
-export async function internal_fetchAcl(
-  resourceInfo: WithServerResourceInfo,
-  options: Partial<
-    typeof internal_defaultFetchOptions
-  > = internal_defaultFetchOptions
-): Promise<WithAcl["internal_acl"]> {
-  if (!hasAccessibleAcl(resourceInfo)) {
-    return {
-      resourceAcl: null,
-      fallbackAcl: null,
-    };
-  }
-  const resourceAcl = await internal_fetchResourceAcl(resourceInfo, options);
-
-  const acl =
-    resourceAcl === null
-      ? {
-          resourceAcl: null,
-          fallbackAcl: await internal_fetchFallbackAcl(resourceInfo, options),
-        }
-      : { resourceAcl: resourceAcl, fallbackAcl: null };
-
-  return acl;
-}
 
 /**
  * @internal
