+++ src/resource/resource.test.ts
@@ -38,14 +38,9 @@ import {
   isPodOwner,
   FetchError,
 } from "./resource";
-import { internal_cloneResource, internal_fetchAcl } from "./resource.internal";
+import { internal_cloneResource } from "./resource.internal";
 
-import {
-  isContainer,
-  isRawData,
-  getContentType,
-  getResourceInfoWithAcl,
-} from "./resource";
+import { isContainer, isRawData, getContentType } from "./resource";
 import {
   WithResourceInfo,
   IriString,
@@ -65,335 +60,6 @@ type MockedFetch = jest.Mock<
   ReturnType<typeof window.fetch>,
   Parameters<typeof window.fetch>
 >;
-
-describe("fetchAcl", () => {
-  it("calls the included fetcher by default", async () => {
-    const mockedFetcher = jest.requireMock("../fetcher.ts") as {
-      fetch: MockedFetch;
-    };
-
-    const mockResourceInfo: WithServerResourceInfo = {
-      internal_resourceInfo: {
-        sourceIri: "https://some.pod/resource",
-        isRawData: false,
-        aclUrl: "https://some.pod/resource.acl",
-        linkedResources: {
-          acl: ["https://some.pod/resource.acl"],
-        },
-      },
-    };
-
-    await internal_fetchAcl(mockResourceInfo);
-
-    expect(mockedFetcher.fetch.mock.calls[0][0]).toEqual(
-      "https://some.pod/resource.acl"
-    );
-  });
-
-  it("does not attempt to fetch ACLs if the fetched Resource does not include a pointer to an ACL file, and sets an appropriate default value.", async () => {
-    const mockFetch = jest.fn(window.fetch);
-
-    const mockResourceInfo: WithServerResourceInfo = {
-      internal_resourceInfo: {
-        sourceIri: "https://some.pod/resource",
-        isRawData: false,
-        linkedResources: {},
-      },
-    };
-
-    const fetchedAcl = await internal_fetchAcl(mockResourceInfo, {
-      fetch: mockFetch,
-    });
-
-    expect(mockFetch.mock.calls).toHaveLength(0);
-    expect(fetchedAcl.resourceAcl).toBeNull();
-    expect(fetchedAcl.fallbackAcl).toBeNull();
-  });
-
-  it("returns null for the Container ACL if the Container's ACL file could not be fetched", async () => {
-    const mockFetch = jest.fn((url) => {
-      const headers =
-        url === "https://some.pod/resource"
-          ? { Link: '<resource.acl>; rel="acl"' }
-          : url === "https://some.pod/"
-          ? { Link: "" }
-          : undefined;
-      return Promise.resolve(
-        mockResponse(undefined, {
-          headers: headers,
-          url: url as string,
-        })
-      );
-    });
-
-    const mockResourceInfo: WithServerResourceInfo = {
-      internal_resourceInfo: {
-        sourceIri: "https://some.pod/resource",
-        isRawData: false,
-        aclUrl: "https://some.pod/resource.acl",
-        linkedResources: {
-          acl: ["https://some.pod/resource.acl"],
-        },
-      },
-    };
-
-    const fetchedAcl = await internal_fetchAcl(mockResourceInfo, {
-      fetch: mockFetch,
-    });
-
-    expect(fetchedAcl).not.toBeNull();
-    expect(fetchedAcl.fallbackAcl).toBeNull();
-    expect(fetchedAcl.resourceAcl?.internal_resourceInfo.sourceIri).toBe(
-      "https://some.pod/resource.acl"
-    );
-  });
-
-  it("returns the fallback ACL even if the Resource's own ACL could not be found", async () => {
-    const mockFetch = jest.fn((url) => {
-      if (url === "https://some.pod/resource.acl") {
-        return Promise.resolve(
-          mockResponse("ACL not found", {
-            status: 404,
-            url: "https://some.pod/resource.acl",
-          })
-        );
-      }
-
-      const headers =
-        url === "https://some.pod/resource"
-          ? { Link: '<resource.acl>; rel="acl"' }
-          : url === "https://some.pod/"
-          ? { Link: '<.acl>; rel="acl"' }
-          : undefined;
-      return Promise.resolve(
-        mockResponse(undefined, {
-          headers: headers,
-          url: url as string,
-        })
-      );
-    });
-
-    const mockResourceInfo: WithServerResourceInfo = {
-      internal_resourceInfo: {
-        sourceIri: "https://some.pod/resource",
-        isRawData: false,
-        aclUrl: "https://some.pod/resource.acl",
-        linkedResources: {
-          acl: ["https://some.pod/resource.acl"],
-        },
-      },
-    };
-
-    const fetchedAcl = await internal_fetchAcl(mockResourceInfo, {
-      fetch: mockFetch,
-    });
-
-    expect(fetchedAcl.resourceAcl).toBeNull();
-    expect(fetchedAcl.fallbackAcl?.internal_resourceInfo.sourceIri).toBe(
-      "https://some.pod/.acl"
-    );
-  });
-});
-
-describe("getResourceInfoWithAcl", () => {
-  it("returns the Resource's own ACL and not its Container's if available", async () => {
-    const mockFetch = jest.fn((url) => {
-      const headers =
-        url === "https://some.pod/resource"
-          ? { Link: '<resource.acl>; rel="acl"' }
-          : url === "https://some.pod/"
-          ? { Link: '<.acl>; rel="acl"' }
-          : undefined;
-      return Promise.resolve(
-        mockResponse(undefined, {
-          headers: headers,
-          url: url as string,
-        })
-      );
-    });
-
-    const fetchedSolidDataset = await getResourceInfoWithAcl(
-      "https://some.pod/resource",
-      { fetch: mockFetch }
-    );
-
-    expect(fetchedSolidDataset.internal_resourceInfo.sourceIri).toBe(
-      "https://some.pod/resource"
-    );
-    expect(
-      fetchedSolidDataset.internal_acl?.resourceAcl?.internal_resourceInfo
-        .sourceIri
-    ).toBe("https://some.pod/resource.acl");
-    expect(fetchedSolidDataset.internal_acl?.fallbackAcl).toBeNull();
-    expect(mockFetch.mock.calls).toHaveLength(2);
-    expect(mockFetch.mock.calls[0][0]).toBe("https://some.pod/resource");
-    expect(mockFetch.mock.calls[1][0]).toBe("https://some.pod/resource.acl");
-  });
-
-  it("returns the Resource's Container's ACL if its own ACL is not available", async () => {
-    const mockFetch = jest.fn((url) => {
-      if (url === "https://some.pod/resource.acl") {
-        return Promise.resolve(new Response("Not found", { status: 404 }));
-      }
-
-      const headers =
-        url === "https://some.pod/resource"
-          ? { Link: '<resource.acl>; rel="acl"' }
-          : url === "https://some.pod/"
-          ? { Link: '<.acl>; rel="acl"' }
-          : undefined;
-      return Promise.resolve(
-        mockResponse(undefined, {
-          headers: headers,
-          url: url as string,
-        })
-      );
-    });
-
-    const fetchedSolidDataset = await getResourceInfoWithAcl(
-      "https://some.pod/resource",
-      { fetch: mockFetch }
-    );
-
-    expect(fetchedSolidDataset.internal_resourceInfo.sourceIri).toBe(
-      "https://some.pod/resource"
-    );
-    expect(fetchedSolidDataset.internal_acl?.resourceAcl).toBeNull();
-    expect(
-      fetchedSolidDataset.internal_acl?.fallbackAcl?.internal_resourceInfo
-        .sourceIri
-    ).toBe("https://some.pod/.acl");
-    expect(mockFetch.mock.calls).toHaveLength(4);
-    expect(mockFetch.mock.calls[0][0]).toBe("https://some.pod/resource");
-    expect(mockFetch.mock.calls[1][0]).toBe("https://some.pod/resource.acl");
-    expect(mockFetch.mock.calls[2][0]).toBe("https://some.pod/");
-    expect(mockFetch.mock.calls[3][0]).toBe("https://some.pod/.acl");
-  });
-
-  it("calls the included fetcher by default", async () => {
-    const mockedFetcher = jest.requireMock("../fetcher.ts") as {
-      fetch: MockedFetch;
-    };
-
-    await getResourceInfoWithAcl("https://some.pod/resource");
-
-    expect(mockedFetcher.fetch.mock.calls).toEqual([
-      [
-        "https://some.pod/resource",
-        {
-          method: "HEAD",
-        },
-      ],
-    ]);
-  });
-
-  it("does not attempt to fetch ACLs if the fetched Resource does not include a pointer to an ACL file, and sets an appropriate default value.", async () => {
-    const mockFetch = jest.fn(window.fetch);
-
-    mockFetch.mockReturnValueOnce(
-      Promise.resolve(
-        mockResponse(undefined, {
-          headers: {
-            Link: "",
-          },
-          url: "https://some.pod/resource",
-        })
-      )
-    );
-
-    const fetchedSolidDataset = await getResourceInfoWithAcl(
-      "https://some.pod/resource",
-      { fetch: mockFetch }
-    );
-
-    expect(mockFetch.mock.calls).toHaveLength(1);
-    expect(fetchedSolidDataset.internal_acl.resourceAcl).toBeNull();
-    expect(fetchedSolidDataset.internal_acl.fallbackAcl).toBeNull();
-  });
-
-  it("returns a meaningful error when the server returns a 403", async () => {
-    const mockFetch = jest
-      .fn(window.fetch)
-      .mockReturnValue(
-        Promise.resolve(new Response("Not allowed", { status: 403 }))
-      );
-
-    const fetchPromise = getResourceInfoWithAcl(
-      "https://arbitrary.pod/resource",
-      {
-        fetch: mockFetch,
-      }
-    );
-
-    await expect(fetchPromise).rejects.toThrow(
-      new Error(
-        "Fetching the metadata of the Resource at `https://arbitrary.pod/resource` failed: `403` `Forbidden`."
-      )
-    );
-  });
-
-  it("returns a meaningful error when the server returns a 404", async () => {
-    const mockFetch = jest
-      .fn(window.fetch)
-      .mockReturnValue(
-        Promise.resolve(new Response("Not found", { status: 404 }))
-      );
-
-    const fetchPromise = getResourceInfoWithAcl(
-      "https://arbitrary.pod/resource",
-      {
-        fetch: mockFetch,
-      }
-    );
-
-    await expect(fetchPromise).rejects.toThrow(
-      new Error(
-        "Fetching the metadata of the Resource at `https://arbitrary.pod/resource` failed: `404` `Not Found`."
-      )
-    );
-  });
-
-  it("includes the status code and status message when a request failed", async () => {
-    const mockFetch = jest.fn(window.fetch).mockReturnValue(
-      Promise.resolve(
-        new Response("I'm a teapot!", {
-          status: 418,
-          statusText: "I'm a teapot!",
-        })
-      )
-    );
-
-    const fetchPromise = getResourceInfoWithAcl(
-      "https://arbitrary.pod/resource",
-      {
-        fetch: mockFetch,
-      }
-    );
-
-    await expect(fetchPromise).rejects.toMatchObject({
-      statusCode: 418,
-      statusText: "I'm a teapot!",
-    });
-  });
-
-  it("does not request the actual data from the server", async () => {
-    const mockFetch = jest
-      .fn(window.fetch)
-      .mockReturnValue(
-        Promise.resolve(
-          mockResponse(undefined, { url: "https://some.pod/resource" })
-        )
-      );
-
-    await getResourceInfoWithAcl("https://some.pod/resource", {
-      fetch: mockFetch,
-    });
-
-    expect(mockFetch.mock.calls).toEqual([
-      ["https://some.pod/resource", { method: "HEAD" }],
-    ]);
-  });
-});
 
 describe("getResourceInfo", () => {
   it("calls the included fetcher by default", async () => {
