+++ src/resource/solidDataset.test.ts
@@ -39,7 +39,6 @@ import {
   getSolidDataset,
   saveSolidDatasetAt,
   saveSolidDatasetInContainer,
-  getSolidDatasetWithAcl,
   createSolidDataset,
   createContainerAt,
   createContainerInContainer,
@@ -351,180 +350,6 @@ describe("getSolidDataset", () => {
     const fetchPromise = getSolidDataset("https://arbitrary.pod/resource", {
       fetch: mockFetch,
     });
-
-    await expect(fetchPromise).rejects.toMatchObject({
-      statusCode: 418,
-      statusText: "I'm a teapot!",
-    });
-  });
-});
-
-describe("getSolidDatasetWithAcl", () => {
-  it("returns the Resource's own ACL and not its Container's if available", async () => {
-    const mockFetch = jest.fn((url) => {
-      const headers =
-        url === "https://some.pod/resource"
-          ? { Link: '<resource.acl>; rel="acl"' }
-          : url === "https://some.pod/"
-          ? { Link: '<.acl>; rel="acl"' }
-          : undefined;
-      return Promise.resolve(
-        mockResponse(undefined, {
-          headers: headers,
-          url: url as string,
-        })
-      );
-    });
-
-    const fetchedSolidDataset = await getSolidDatasetWithAcl(
-      "https://some.pod/resource",
-      { fetch: mockFetch }
-    );
-
-    expect(fetchedSolidDataset.internal_resourceInfo.sourceIri).toBe(
-      "https://some.pod/resource"
-    );
-    expect(
-      fetchedSolidDataset.internal_acl?.resourceAcl?.internal_resourceInfo
-        .sourceIri
-    ).toBe("https://some.pod/resource.acl");
-    expect(fetchedSolidDataset.internal_acl?.fallbackAcl).toBeNull();
-    expect(mockFetch.mock.calls).toHaveLength(2);
-    expect(mockFetch.mock.calls[0][0]).toBe("https://some.pod/resource");
-    expect(mockFetch.mock.calls[1][0]).toBe("https://some.pod/resource.acl");
-  });
-
-  it("returns the Resource's Container's ACL if its own ACL is not available", async () => {
-    const mockFetch = jest.fn((url) => {
-      if (url === "https://some.pod/resource.acl") {
-        return Promise.resolve(new Response("Not found", { status: 404 }));
-      }
-
-      const headers =
-        url === "https://some.pod/resource"
-          ? { Link: '<resource.acl>; rel="acl"' }
-          : url === "https://some.pod/"
-          ? { Link: '<.acl>; rel="acl"' }
-          : undefined;
-      return Promise.resolve(
-        mockResponse(undefined, {
-          headers: headers,
-          url: url as string,
-        })
-      );
-    });
-
-    const fetchedSolidDataset = await getSolidDatasetWithAcl(
-      "https://some.pod/resource",
-      { fetch: mockFetch }
-    );
-
-    expect(fetchedSolidDataset.internal_resourceInfo.sourceIri).toBe(
-      "https://some.pod/resource"
-    );
-    expect(fetchedSolidDataset.internal_acl?.resourceAcl).toBeNull();
-    expect(
-      fetchedSolidDataset.internal_acl?.fallbackAcl?.internal_resourceInfo
-        .sourceIri
-    ).toBe("https://some.pod/.acl");
-    expect(mockFetch.mock.calls).toHaveLength(4);
-    expect(mockFetch.mock.calls[0][0]).toBe("https://some.pod/resource");
-    expect(mockFetch.mock.calls[1][0]).toBe("https://some.pod/resource.acl");
-    expect(mockFetch.mock.calls[2][0]).toBe("https://some.pod/");
-    expect(mockFetch.mock.calls[3][0]).toBe("https://some.pod/.acl");
-  });
-
-  it("calls the included fetcher by default", async () => {
-    const mockedFetcher = jest.requireMock("../fetcher.ts") as {
-      fetch: jest.Mock<
-        ReturnType<typeof window.fetch>,
-        [RequestInfo, RequestInit?]
-      >;
-    };
-
-    await getSolidDatasetWithAcl("https://some.pod/resource");
-
-    expect(mockedFetcher.fetch.mock.calls[0][0]).toEqual(
-      "https://some.pod/resource"
-    );
-  });
-
-  it("does not attempt to fetch ACLs if the fetched Resource does not include a pointer to an ACL file, and sets an appropriate default value.", async () => {
-    const mockFetch = jest.fn(window.fetch);
-
-    mockFetch.mockReturnValueOnce(
-      Promise.resolve(
-        mockResponse(undefined, {
-          headers: {
-            Link: "",
-          },
-          url: "https://some.pod/resource",
-        })
-      )
-    );
-
-    const fetchedSolidDataset = await getSolidDatasetWithAcl(
-      "https://some.pod/resource",
-      { fetch: mockFetch }
-    );
-
-    expect(mockFetch.mock.calls).toHaveLength(1);
-    expect(fetchedSolidDataset.internal_acl.resourceAcl).toBeNull();
-    expect(fetchedSolidDataset.internal_acl.fallbackAcl).toBeNull();
-  });
-
-  it("returns a meaningful error when the server returns a 403", async () => {
-    const mockFetch = jest
-      .fn(window.fetch)
-      .mockReturnValue(
-        Promise.resolve(new Response("Not allowed", { status: 403 }))
-      );
-
-    const fetchPromise = getSolidDatasetWithAcl("https://some.pod/resource", {
-      fetch: mockFetch,
-    });
-
-    await expect(fetchPromise).rejects.toThrow(
-      new Error(
-        "Fetching the Resource at `https://some.pod/resource` failed: `403` `Forbidden`."
-      )
-    );
-  });
-
-  it("returns a meaningful error when the server returns a 404", async () => {
-    const mockFetch = jest
-      .fn(window.fetch)
-      .mockReturnValue(
-        Promise.resolve(new Response("Not found", { status: 404 }))
-      );
-
-    const fetchPromise = getSolidDatasetWithAcl("https://some.pod/resource", {
-      fetch: mockFetch,
-    });
-
-    await expect(fetchPromise).rejects.toThrow(
-      new Error(
-        "Fetching the Resource at `https://some.pod/resource` failed: `404` `Not Found`."
-      )
-    );
-  });
-
-  it("includes the status code and status message when a request failed", async () => {
-    const mockFetch = jest.fn(window.fetch).mockReturnValue(
-      Promise.resolve(
-        new Response("I'm a teapot!", {
-          status: 418,
-          statusText: "I'm a teapot!",
-        })
-      )
-    );
-
-    const fetchPromise = getSolidDatasetWithAcl(
-      "https://arbitrary.pod/resource",
-      {
-        fetch: mockFetch,
-      }
-    );
 
     await expect(fetchPromise).rejects.toMatchObject({
       statusCode: 418,
